backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
firebase-admin==6.5.0
reportlab==4.2.2
pytest==8.3.2
gunicorn==22.0.0
